+++ tests/test_sfsifiles.py
@@ -421,12 +421,25 @@ def test_can_set_more_than_two_positional_args():
     assert cus.p3 == 5
 
 
+def test_save_and_load_soil_w_diff_liq_mass_density():
+    sl = models.Soil(liq_mass_density=1.0)
+    sl.e_curr = 0.65
+    sl.specific_gravity = 2.65
+    ecp_output = sm.Output()
+    ecp_output.add_to_dict(sl)
+    p_str = json.dumps(ecp_output.to_dict(), indent=4)
+    print(p_str)
+    objs = sm.loads_json(p_str)
+
+
+
 if __name__ == '__main__':
     # test_load_json()
     # test_save_and_load_wall_building()
     # test_save_and_load_building()
     # test_load_and_save_structure()
-    test_can_load_then_save_and_load_custom_ecp_w_custom_obj()
+    # test_save_and_load_soil_w_diff_liq_mass_density()
+    test_load_olf_file_w_frame_w_hinges()
     # test_load_json()
     # test_full_save_and_load()
     # test_save_and_load_soil_profile()